+++ examples/color.rs
@@ -13,7 +13,7 @@ use embedded_graphics_simulator::{
 };
 use embedded_menu::{
     interaction::simulator::Simulator,
-    items::{select::SelectValue, Select},
+    items::{menu_item::SelectValue, MenuItem},
     selection_indicator::style::rectangle::Rectangle as RectangleIndicator,
     theme::Theme,
     Menu, MenuStyle,
@@ -73,10 +73,10 @@ fn main() -> Result<(), core::convert::Infallible> {
                 esc_value: (),
             }),
     )
-    .add_item(Select::new("Foo", ">"))
-    .add_item(Select::new("Check this", false))
-    .add_item(Select::new("Check this", TestEnum::A))
-    .add_item(Select::new("Check this too", false))
+    .add_item(MenuItem::new("Foo", ">"))
+    .add_item(MenuItem::new("Check this", false))
+    .add_item(MenuItem::new("Check this", TestEnum::A))
+    .add_item(MenuItem::new("Check this too", false))
     .build();
 
     let output_settings = OutputSettingsBuilder::new().scale(4).build();