+++ src/items/mod.rs
@@ -1,8 +1,8 @@
+pub mod menu_item;
 mod section_title;
-pub mod select;
 
+pub use menu_item::MenuItem;
 pub use section_title::SectionTitle;
-pub use select::Select;
 
 use crate::{
     interaction::InputAdapterSource,